pyproject.toml
[tool.poetry]
name = "HotelBooking"
version = "0.1.0"
description = ""
authors = ["Budnikov Aleksey <budnikovAV4@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.109.2"}
datetime = "^5.4"
alembic = "^1.13.1"
sqlalchemy = "^2.0.25"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
install = "^1.3.5"
pydantic-settings = "^2.1.0"
pydantic = "^2.6.1"
passlib = "^1.7.4"
bcrypt = "^4.1.2"
python-jose = "^3.3.0"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
celery = "^5.3.6"
flower = "^2.0.1"
pillow = "^10.2.0"
loguru = "^0.7.2"
sqladmin = {extras = ["full"], version = "^0.16.1"}
pytest = "7.2.2"
pytest-asyncio = "0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
